pyproject.toml
[project]
name = "spotify-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.15.0",
    "python-dotenv>=1.1.1",
    "spotipy>=2.25.1",
    "uvicorn>=0.37.0",
]
